Python Trace and Save Python State are not compatible
* Both "Save Python State" and "Trace" rely on vtkSMTrace,
that does not allow to be started twice at a time.

So saving a Python state while running the Trace leads
no output file and error message.
It also means that Auto Save Python state can not work while
Trace is active. So we choose to disable auto-save when
starting trace, and re enable it after stop.

See #22735