CAVE: Fix role bindings for valuator events
Calling GetValuatorRole(eventName) has never worked because valuator
events are configured with an id, which is an index, but are delivered
with all ids/indices of valuator as a single event.

Since the relationship between role:event isn't 1:1, remove the
GetValuatorRole method. Instead, provide a method to get the id/index
associated with given a role name.