preset colormaps of both halves of fast
This commit adds two colormaps, Fast (Reds) and Fast (Blues).
These colormaps may be useful for creating plots of values that
are known to be either all-positive (von Mises stress) or
all-negative values (change in temperature for a cooling object).
The colormaps were created by copying the XML data for the Fast
colormap, and reassigning control points using affine mappings of
Reds: [0.5, 1] -> [0, 1] and Blues [0, 0.5] -> [0, 1].