Do not force AnimationScene to be part of python states
When using the `source_set` arguments, listing the
desired proxies, we should not force AnimationScene to
be traced.
Do it only when the argument is empty, as we do for Sources and Views.

This was impacting Catalyst script, were RenderView ends
to be there even without Screenshots extractors

Was introduce in !6352

(cherry picked from commit bf50e57c1cc9ebef730e89bfa04e4d38c38f7eea)