WrapClientServer: fix export macro usage
This is an unconditional export. `VTK_EXPORT` is sensitive to whether
VTK is built as static or shared; these uses should not be tied to that
as they are separate decisions.

Additionally, use `VTK_ABI_HIDDEN` to hide symbols which are only called
from the entry point; there's no need to call them from outside of the
hosting library. Other functions may be marked `static` as they are not
used outside of the file in question.

(cherry picked from commit 9d06f9b025b2edf7934d2f3cd41487e9a01f8736)