Configure a SystemScope QSettings path
QSettings has a fallback mechanism: when a key is not found
in UserScope file, it looks for a SystemScope one.

Hook the system paths to look for desired ParaView install paths,
instead of the system paths checked by default by Qt.

Note that the .ini file still needs to be under a "Organization" directory
(aka ParaView in our case)

https://doc.qt.io/qt-5/qsettings.html#platform-specific-notes