Fix regression on CreateTexture python function
When using the CreateTexture function to read from a file,
the filename was not being set correctly.
This issue occurred because the RegisterTextureProxy was not considered,
as the _getPyProxy had already registered the proxy.
The TrivialProducerKey and Mode proxy properties are now set in the
Register functions.

(cherry picked from commit af3c51cc349d9f9f7dc114a1a42c089634e8463c)