Reading proxypropety settings should set property and not modify domain
When reading a proxy property (such as clip type: plane, cylinder etc)
one is marked as Selected: it should be set as the current value for
the property and not as the domain default.

With previous code, the property was not properly updated when creating the proxy.
Instead it was updated when using "Reset to App default" even if this settings is
a "User default" and not an "App default".

See SettingsProxyProperty XML test