vtkSMSettings should start with a writeable collection
In --disable-registry mode, no collections are added
under UserPriority, but an app one is created. (see vtkInitializationHelper)
Then "SetXXXSettings" are applied on an Application collection.
This is an issue because one can still play with save/restore user/app
default from the GUI, which leads to different behavior than
doing so without --dr.