Spy Plot Reader: pass properties to readers after they are created
The DownConvertVolumeFraction and GenerateMarkers properties are
passed from the vtkSpyPlotReader to vtkSpyPlotUniReader children
objects it creates via the setters SetDownConvertVolumeFraction and
SetGenerateMarkers, respectively. However, these children objects are
often created during the RequestData phase which occurs after the
property is set on the vtkSpyPlotReader. As a result, the properties
of the vtkSpyPlotUniReader objects were not getting set the up-to-date
property values. Fixed that by forwarding these property values just
after the child reader objects are created in
vtkSpyPlotReaderMap::GetReader.

Fixes #22571.