Keep track of `vtkSMProxyProperty` defaults
Previously, proxy properties would only assume they were default if
there were contained no proxies. However, `vtkSMProxyListDomain` typically
has default proxies that it sets for the proxy property in
`vtkSMProxyListDomain::SetDefaultValues()` (specifically, if there
is only one value in the list or a `default` explicitly specified in
the XML, that is used as the default).

In a similar vein to other proxies such as `vtkSMDoubleVectorProperty`,
we now keep track of the default proxies within `vtkSMProxyProperty`.
When `vtkSMProxyListDomain::SetDefaultValues()` is called and sets
a default value on a proxy property, it now also triggers the proxy
property to store the current values it has as its defaults.

`IsValueDefault()` for the proxy property now checks if the current
proxies match the internally stored defaults, and it also recursively
checks that the properties of each proxy in a proxy property are default
as well. This will likely result in a reduction of code in settings files,
since fewer proxy properties will be written out compared to before (the
previous behavior was to always write out every non-empty proxy property).

There is also a new function on a proxy property called
`IsValueDefaultNonRecursive()`, which verifies that all proxies and ports of a
proxy property match the default, but it does not check that the properties of
each proxy in a proxy property are default. This is used in Python state files
to determine whether a proxy property needs to be written out.

This results in a reduction of the content in Python state files when the trace
is set to only write out modified properties.

For a simple example where I created a wavelet, a contour, and a
couple of extractors, the Python state file size was reduced from
215 lines to 199 lines. For example, the `renderView` was changed
from this:

```python
renderView1.Set(
    AxesGrid='Grid Axes 3D Actor',
    StereoType=0,
    CameraPosition=[0.0, -0.3340263366699219, 66.1845775207595],
    CameraFocalPoint=[0.0, -0.3340263366699219, 0.0],
    CameraFocalDisk=1.0,
    CameraParallelScale=17.129829154436734,
    LegendGrid='Legend Grid Actor',
    PolarGrid='Polar Grid Actor',
)
```

to this (three default proxy properties were removed):

```python
renderView1.Set(
    StereoType=0,
    CameraPosition=[0.0, -0.3340263366699219, 66.1845775207595],
    CameraFocalPoint=[0.0, -0.3340263366699219, 0.0],
    CameraFocalDisk=1.0,
    CameraParallelScale=17.129829154436734,
)
```

Several other default proxy properties were likewise removed from
other functions. For example, `PointMergeMethod='Uniform Binning'`
was removed from the contour settings, the following were all
removed from the geometry representation:

```python
    TextureTransform='Transform2',
    OSPRayScaleFunction='Piecewise Function',
    GlyphType='Arrow',
    ScaleTransferFunction='Piecewise Function',
    OpacityTransferFunction='Piecewise Function',
    DataAxesGrid='Grid Axes Representation',
    PolarAxes='Polar Axes Representation',
```

And `Trigger = 'Time Step'` was removed from the extractors.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>