Change "NameOfLastPresetApplied" to be a property
If `NameOfLastPresetApplied` is a property, it gets saved in `.pvsm`
state files. This can be helpful if users save a `.pvsm` file, load it,
then save a Python state file - the name will be remembered.

We blacklist it in the Python trace because we set it explicitly. Maybe
this means that if a user loads a Python state file with a LUT applied,
and then write it out again as a Python state file, that it may be
forgotten (and the whole RGBPoints would get dumped), but maybe that's
something we can fix in the future if someone actually encounters it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>