Change where status of example visualization loading is shown
Show the message in the status bar of the main window rather than in a
popup dialog box.

This change is motivated by behavior on macOS with Qt6 where the
window always showed an OK button that needed to be clicked to make
the dialog disappear. There seemed to be no way to remove that button.
The status bar is perfect for this kind of message anyway.