Skip writing `None` strings in Python state
In the state file, the representation used to set display properties that
were just the `None` string (obtained from the `none_string` in the XML
of the array list domain). For example, the following would show up in
the representation properties:

```python
    SelectTangentArray='None',
    SelectTCoordArray='None',
    SelectOrientationVectors='None',
```

However, if we leave them out of the state file (so they remain as their
default of an empty string), they get changed to `None` automatically anyways
because of the `none_string` logic.

Thus, we should just leave them out of the Python state file. This commit
removes them.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>