Increase width of tooltips generated by pqPropertyWidget
Resolves issue #17427 by increasing the width of tooltips on most user
interface elements, and removing the text justification that was
previously applied. Since QToolTip doesn't provide a way to set its
width, and the rendered rich text ignores any width style properties
applied in the HTML tags, the compromise was to display text as a single
non-wrapped line for short text, and display long text in a one-celled
table with a fixed length that forces the text to wrap.