Qt6: Start container with LC_ALL=en_US.UTF-8
The container locally starts with LANG=C.UTF-8.

This is needed so that calls like below have the thousands seperator,
which various tests expect

auto l = QLocale::system();
l.toString(dinfo->GetNumberOfElements(item.first))

I checked what the LANG was in fedora35, and it was still the same, which
makes me question how in the world Qt5 was even working in the first place.