Fix multiline strings in Python trace output
Some recent indentation changes would result in multiline strings in
Python trace being set like this:

```python
programmableFilter1.Set(
    Script="""print("a")
    print("b")""",
)
```

This would cause errors when loading it, because the indentation makes
the Python script invalid. This MR fixes it so that it is now like this:

```python
programmableFilter1.Set(
    Script="""print("a")
print("b")""",
)
```

Which is similar to the previous behavior. Any properties containing
multiline strings are also moved to the back of the `Set()` function
because they are easier to read that way.

A test was added that verifies this fix, and ensures it does not
happen again.

backport: release

Fixes: #22989

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>