FIX: vtkPVMetaSliceDataSet was handling mtime correctly. Fixed.
vtkPVMetaSliceDataSet was relying only on the selected filter to
compute mtime. However, it does not set the implicit function
on the filter until filter execution. This means that the mtime
of the implicit function(s) it uses is not taken into account
unless the filter executes. This is compounded further if it
is executing in parallel and some of the ranks have empty
composite datasets (which means that RequestData is never called).
This fix directly takes into account the mtime of the implicit
functions.