color.py: GenerateRGBPoints() returns empty list if preset_name can't be found
If a `preset_name` argument is provided but can't be found in the list
of presets, GenerateRGBPoints() returns an empty list of values
instead of the list of values from the default. This should prevent
confusion about why invalid preset names return a list of RGB values,
and it enables some error checking based on the returned value of
GenerateRGBPoints().