Disable Plot3DReader.xml when pdal is on
When pdal is loaded there is exactly one reader
capable of reading files with bin in the name.
Previously no reader could read it, and a reader
needed to be selected.

Also set it to use force serial since it can lose focus
because of popup windows.

Additionally, for some reason openpmd intervenes with this test
and causes errors when it's not installed, so make sure it's only
enabled when the package exists.