annotation.py: convert evaluated result to string
vtkPythonAnnotation::SetComputedAnnotationValue() takes a string in
all cases, and the Python formatting string could not convert
arbitrary Python values to a string in all cases, such as the
`time_range` tuple. Wrapping the `result` variable in `str()` solves
this problem, and should set the annotation value to what one would
see as the output of a Python interpreter.