vtkPythonCalculator: treat as a time source
Imbue vtkPythonCalculator with a TIME_RANGE() so that the pipeline
executive will update it when the current time changes. This is to
ensure that Python Calculator expressions with `current_time` get
updated when the time changes. Without it, the executive will not
update the filter because it will be in an unmodified state.