Use a numerical sort for pqArraySelectionWidget items
Perform actual numerical sort in table when using a numerical column, instead of the confusing
alphabetical sort (where `100` used to be put before `4`).

In the 2-columns mode of the pqArraySelectionWidget, one column is an "id" (typically in the IOSS readers).
This id is a numerical value stored as a string (comes from a 2 components string vector property),
so it is hard to modify this type.

Updating Qt code to handle this case.
While `QCollator` seems to exist for that purpose, its numerical support
depends on how Qt was built (https://doc.qt.io/qt-6.9/qcollator.html#posix-fallback-implementation)
so we implement it by ourself