default colormap "Fast": use the correct "Lab" space instead of "Diverging"
When the default color map was changed to "Fast", the default color
space was not updated to "Lab" but instead was left as
"Diverging". The Fast color map with "Diverging" color space is much
different from the color map with the correct "Lab" space, and has a
bright white offset from the middle of the color map.