Utilities/VisItBridge: Fix crash with SILO reader
There was a crash when reading in some SILO files (see #22373
and https://github.com/LLNL/Silo/issues/445). After chasing things
down, the reader was attempting to read names in the metadata after
turning of the loading of names in the read. Fixed this issue by
adding some SILO file read flags to get the appropriate information.

* Kenneth Moreland: fix-silo-reader (paraview/visitbridge!119)
* Kenneth Moreland: error-reporting (paraview/visitbridge!117)