vtkPVRenderView: enable/disable Camera Orientation Widget interaction
The Camera Orientation Widget was always enabled, even when not
visible.  This meant that in a vtkPVRenderView's 2D interaction mode,
one could rotate the scene by interacting with the invisible Camera
Orientation Widget, i.e., click and dragging in the area in which the
Camera Orientation Widget would appear if it were visible. Fixed this
behavior by enabling the widget only when it is visible.