gitlab-ci: support a mechanism to vet new CI machines
When deploying a new ParaView CI machine, it should first be deployed
with the `__paraview` tag. Then a pipeline can be created on ParaView
itself to run on said new machines to verify that they are running CI
reliably. Once verified, the machine's runners may be switched back to
using the `paraview` project tag.