CAVEInteraction: Add a world scale value to the panel
Allow setting a world scaling factor which is applied to the
ModelTransformMatrix when updated. Also add a navigation event
observer to the CAVE config panel that extracts the scaling
value, so that the UI always reflects the correct current scale.

Extracting/updating the scale factor from/to the ModelTransformMatrix
are added as static convenience methods to vtkSMVRInteractorStyleProxy.

Add a new example Python interactor style that performs scaling, so that
the connection between scaling with an interactor style and scaling
using the panel QLineEdit can be observed and verified.