Fix compilation against the new VTK statistics algorithms.
This fixes old vtkSciVizStatistics and subclasses plus tests
which involve them.
Internally, it uses the new vtkGenerateStatistics and provides a
way to create statistical models via filters as opposed to the
statistics view.

The PartialArrayInLineChart test XML required changes because the
output model is organized differently in the new version (the model
hierarchy is deeper; also multiple requests are processed when more
than 2 input variables are selected).